+++ changelog.md
@@ -12,6 +12,8 @@
 
   * Add `flushTBQueue` to `Control.Concurrent.STM.TBQueue` (gh-1)
 
+  * Add `signalTSemN` operations (gh-5)
+
 
 ## 2.4.4.1  *Dec 2015*